Adds assertions to execute() methods
There are no assertions used anywhere.

Assertions help check for assumptions made about the code and keep
it error free.

Use assertions to ensure that task is added to task list when
method execute() is called.

Helps check that the task is actually added to the task list and
ensure that addTask and TaskList are functioning as expected.